+++ lib/stardag/src/stardag/testing/modal/__init__.py
@@ -1,0 +1,1 @@
+"""Modal integration test utilities (internal)."""